update to u16

+++ world-chain-builder/crates/world/node/src/args.rs
@@ -11,7 +11,7 @@ pub struct WorldChainArgs {
 
     /// Sets the number of allowed PBH transactions per month
     #[arg(long = "builder.num_pbh_txs", default_value = "30")]
-    pub num_pbh_txs: u8,
+    pub num_pbh_txs: u16,
 
     /// Sets the max blockspace reserved for verified transactions. If there are not enough
     /// verified transactions to fill the capacity, the remaining blockspace will be filled with
